robohash/scripts/requirements.txt
Pillow==7.0.0
webp==0.1.6